+++ cognition/src/main/scala/app/models/Models.scala
@@ -1,4 +1,4 @@
 package models
 
-class World(val id: Int, var randomNumber: Int)
+class World(val id: Long, var randomNumber: Int)
 class Fortune(val id: Long, var message: String)
